Move certbot-dns-digitalocean to _internal structure
