+++ certbot-dns-digitalocean/certbot_dns_digitalocean/_internal/__init__.py
@@ -1,0 +1,1 @@
+Internal implementation of `~certbot_dns_digitalocean.dns_digitalocean` plugin.
